[CELEBORN-530][REFACTOR] Move buffer manager and memory manager to worker module.
