+++ common/src/test/java/org/apache/celeborn/common/network/TestUtils.java
@@ -18,7 +18,6 @@
 package org.apache.celeborn.common.network;
 
 import java.net.InetAddress;
-import java.util.concurrent.Callable;
 
 public class TestUtils {
   public static String getLocalHost() {
@@ -26,18 +25,6 @@ public class TestUtils {
       return InetAddress.getLocalHost().getHostAddress();
     } catch (Exception e) {
       throw new RuntimeException(e);
-    }
-  }
-
-  public static void timeOutOrMeetCondition(Callable<Boolean> callable) throws Exception {
-    int timeout = 10000; // 10s
-    while (true) {
-      if (callable.call() || timeout < 0) {
-        break;
-      }
-
-      timeout = timeout - 100;
-      Thread.sleep(100);
     }
   }
 }
