+++ worker/src/main/java/org/apache/celeborn/service/deploy/worker/memory/RecyclableBuffer.java
@@ -15,31 +15,22 @@
  * limitations under the License.
  */
 
-package org.apache.celeborn.common.network.server.memory;
+package org.apache.celeborn.service.deploy.worker.memory;
 
-import static com.google.common.base.Preconditions.checkArgument;
+import io.netty.buffer.ByteBuf;
 
-public class ReadBufferRequest {
-  private final int number;
-  private final int bufferSize;
-  private final ReadBufferListener readBufferListener;
+public class RecyclableBuffer {
 
-  public ReadBufferRequest(int number, int bufferSize, ReadBufferListener readBufferListener) {
-    checkArgument(number > 0);
-    this.number = number;
-    this.bufferSize = bufferSize;
-    this.readBufferListener = readBufferListener;
+  public final ByteBuf byteBuf;
+
+  public final BufferRecycler bufferRecycler;
+
+  public RecyclableBuffer(ByteBuf byteBuf, BufferRecycler bufferRecycler) {
+    this.byteBuf = byteBuf;
+    this.bufferRecycler = bufferRecycler;
   }
 
-  public int getNumber() {
-    return number;
-  }
-
-  public int getBufferSize() {
-    return bufferSize;
-  }
-
-  public ReadBufferListener getBufferListener() {
-    return readBufferListener;
+  public void recycle() {
+    bufferRecycler.recycle(byteBuf);
   }
 }
